[CHORE/#11] 테마 파일 수정 - 파일 위치 수정 - 코드 위치 수정
