+++ app/src/main/java/com/sopt/shinmungo/core/designsystem/theme/Theme.kt
@@ -4,18 +4,24 @@ import android.app.Activity
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
+import androidx.compose.runtime.ReadOnlyComposable
 import androidx.compose.runtime.SideEffect
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.toArgb
 import androidx.compose.ui.platform.LocalView
 import androidx.core.view.WindowCompat
-import com.sopt.shinmungo.core.designsystem.color.Gray13
-import com.sopt.shinmungo.core.designsystem.color.LocalColorProvider
-import com.sopt.shinmungo.core.designsystem.color.ShinMunGoColors
-import com.sopt.shinmungo.core.designsystem.color.defaultShinMunGoColors
-import com.sopt.shinmungo.core.designsystem.type.LocalTypographyProvider
-import com.sopt.shinmungo.core.designsystem.type.ShinMunGoTypography
-import com.sopt.shinmungo.core.designsystem.type.defaultShinMunGoTypography
+
+object ShinMunGoTheme {
+    val color: ShinMunGoColors
+        @Composable
+        @ReadOnlyComposable
+        get() = LocalColorProvider.current
+
+    val typography: ShinMunGoTypography
+        @Composable
+        @ReadOnlyComposable
+        get() = LocalTypographyProvider.current
+}
 
 @Composable
 fun ShinMunGoTheme(
